+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # svelte-streamable changelog
 
+## 1.1.0
+
+* Call callback with `undefined` data on int.
+* Fallback to default value every time when result data is `undefined`.
+* Remove superfluous logging and event listeners.
+* Few improvements in README.
+* Improved types.
+
 ## 1.0.0
 
 * First release